moved more files around into fitting folders and begun the cog process. currently syncing the commands does not work :(
