+++ old/kittyleague.py
